utils/requirements.txt
matplotlib
opencv-python
pillow
numpy